Improvement/upload fallback

+++ src/components/Confirm/index.js
@@ -36,34 +36,24 @@ const Previews = ({capture, step, retakeAction, confirmAction} ) =>  {
   )
 }
 
-const getCapture = (captures) => captures[0]
-
-class Confirm extends Component {
-  render () {
-    const {
+const Confirm = ({
       step,
       method,
-      documentCaptures,
-      faceCaptures,
+      validCaptures,
       actions: {
         deleteCaptures,
         confirmCapture
       }
-    } = this.props
+    }) => {
 
-    const captures = {
-      'document': documentCaptures,
-      'face': faceCaptures
-    }[method]
-    const capture = getCapture(captures)
+  const capture = validCaptures[method][0]
 
-    return <Previews
-      capture={capture}
-      retakeAction={() => deleteCaptures({method})}
-      confirmAction={() => confirmCapture({method, id: capture.id})}
-      step={step}
-    />
-  }
+  return <Previews
+    capture={capture}
+    retakeAction={() => deleteCaptures({method})}
+    confirmAction={() => confirmCapture({method, id: capture.id})}
+    step={step}
+  />
 }
 
 export default Confirm
